Allow moving window perpendicular to top-level container
This creates a new top-level container. A very convenient feature taken
from i3.

Now, the 'root' of the layout is distinct from the top-level node. I
might make them distinct classes in the future.

The behavior of this WM now mimics the behavior of i3 almost perfectly.
There is a very slight difference in the way this feature behaves in
some situations, but it's pretty esoteric.