Bring windows to front when switching focus mode
This makes all floating windows come to the front when switching to that
focus mode, and disappear when going back to tiling mode. This is
probably the best way of handling the lack of "always-on-top" mode.

I'm not too happy with this implementation, because it's slow and
disorienting if you have even a moderate number of windows. It can be
optimized, or even replaced with a minimize/unminimize solution.